requirements.txt
requests
SQLALchemy
sqlalchemy_utils
flask
flask-sqlalchemy
flask-bootstrap
flask-iniconfig
flask-nav
Flask-WTF
PyYAML
WTForms-Alchemy
rauth
-e git+https://github.com/tarekziade/smwogger#egg=smwogger
